+++ frontend/src/backend/wails/openFile.ts
@@ -1,5 +1,5 @@
-import { base64Decode } from "utils/encoding/base64";
-import { FilesStorageKey, OpenFileOptions } from "../internal";
+import { OpenFileOptions } from "../internal";
+import { fileRefToOpenFile } from "./fileRefToOpenFile";
 import { OpenFile } from "./go/main/App";
 import { files } from "./go/models";
 
@@ -8,12 +8,7 @@ export const openFile = async ({ multiple = false, done }: OpenFileOptions) => {
   try {
     files = await OpenFile(multiple);
     if (files?.length) {
-      done(
-        files.map((f) => ({
-          file: new File([base64Decode(f.$$blob64)], f.name),
-          key: f.$$id as FilesStorageKey,
-        })),
-      );
+      done(files.map(fileRefToOpenFile));
     }
   } catch {}
 };